+++ src/Telemetry.Serilog/TelemetryLogEventEnricher.cs
@@ -1,4 +1,6 @@
-﻿using Serilog.Core;
+﻿using Byndyusoft.AspNetCore.Mvc.Telemetry.Data;
+using Byndyusoft.AspNetCore.Mvc.Telemetry.Logging;
+using Serilog.Core;
 using Serilog.Events;
 
 namespace Byndyusoft.AspNetCore.Mvc.Telemetry.Serilog
