+++ src/Telemetry/Data/TelemetryInfoItem.cs
@@ -1,0 +1,15 @@
+﻿namespace Byndyusoft.AspNetCore.Mvc.Telemetry.Data
+{
+    public class TelemetryInfoItem
+    {
+        public TelemetryInfoItem(string key, object? value)
+        {
+            Key = key;
+            Value = value;
+        }
+
+        public string Key { get; }
+
+        public object? Value { get; }
+    }
+}
