+++ src/Telemetry/Definitions/DefaultTelemetryEventNames.cs
@@ -1,0 +1,7 @@
+﻿namespace Byndyusoft.AspNetCore.Mvc.Telemetry.Definitions
+{
+    public static class DefaultTelemetryEventNames
+    {
+        public static string Initialization => "TelemetryRouter.Initialization";
+    }
+}
