+++ src/Telemetry/Definitions/StaticTelemetryUniqueNames.cs
@@ -1,0 +1,7 @@
+﻿namespace Byndyusoft.AspNetCore.Mvc.Telemetry.Definitions
+{
+    public static class StaticTelemetryUniqueNames
+    {
+        public static string BuildConfiguration => "Static.BuildConfiguration";
+    }
+}
