+++ src/Telemetry/Definitions/TelemetryWriterUniqueNames.cs
@@ -1,0 +1,9 @@
+﻿namespace Byndyusoft.AspNetCore.Mvc.Telemetry.Definitions
+{
+    public static class TelemetryWriterUniqueNames
+    {
+        public static string Log => "Log";
+
+        public static string LogPropertyAccessor => "LogPropertyAccessor";
+    }
+}
