+++ src/Telemetry/ITelemetryRouter.cs
@@ -1,0 +1,9 @@
+﻿using Byndyusoft.AspNetCore.Mvc.Telemetry.Data;
+
+namespace Byndyusoft.AspNetCore.Mvc.Telemetry
+{
+    public interface ITelemetryRouter
+    {
+        void ProcessTelemetryEvent(TelemetryEvent telemetryEvent);
+    }
+}
