+++ src/Telemetry/Providers/Interface/IStaticTelemetryDataProvider.cs
@@ -1,0 +1,9 @@
+﻿using Byndyusoft.AspNetCore.Mvc.Telemetry.Data;
+
+namespace Byndyusoft.AspNetCore.Mvc.Telemetry.Providers.Interface
+{
+    public interface IStaticTelemetryDataProvider
+    {
+        TelemetryInfo[] GetTelemetryData();
+    }
+}
