+++ src/Telemetry/Writers/Interfaces/ITelemetryWriter.cs
@@ -1,0 +1,11 @@
+﻿using Byndyusoft.AspNetCore.Mvc.Telemetry.Data;
+
+namespace Byndyusoft.AspNetCore.Mvc.Telemetry.Writers.Interfaces
+{
+    public interface ITelemetryWriter
+    {
+        string WriterUniqueName { get; }
+
+        void Write(TelemetryInfo[] telemetryInfos, bool isStaticData);
+    }
+}
